+++ src/utils/common/exec.c
@@ -259,8 +259,8 @@ static void print_section_dig(const knot_rrset_t **rrsets,
 
 	for (size_t i = 0; i < count; i++) {
 		const knot_rrset_t *rrset = rrsets[i];
-		size_t rrset_rdata_count = knot_rrset_rr_count(rrset);
-		for (size_t j = 0; j < rrset_rdata_count; j++) {
+		uint16_t rrset_rdata_count = knot_rrset_rr_count(rrset);
+		for (uint16_t j = 0; j < rrset_rdata_count; j++) {
 			while (knot_rrset_txt_dump_data(rrset, j, buf, buflen,
 			                                &(style->style)) < 0) {
 				buflen += 4096;
@@ -303,8 +303,8 @@ static void print_section_host(const knot_rrset_t **rrsets,
 		}
 		descr = knot_lookup_by_id(rtypes, rrset->type);
 
-		size_t rrset_rdata_count = knot_rrset_rr_count(rrset);
-		for (size_t j = 0; j < rrset_rdata_count; j++) {
+		uint16_t rrset_rdata_count = knot_rrset_rr_count(rrset);
+		for (uint16_t j = 0; j < rrset_rdata_count; j++) {
 			if (rrset->type == KNOT_RRTYPE_CNAME &&
 			    style->hide_cname) {
 				continue;
